+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 ## [WIP] 0.1.1 / 2017-July-??
 
-- [TODO] Stop streaming with error message on exception
+- Stop streaming with error message on exception
   - HTTP Kit
   - Manifold
   - Immutant
-  - Generic
 
 
 ## 0.1.0 / 2017-July-07